+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the tool_vault plugin will be documented in this file.
 ### Fixed
 - Prevent curl from sending Authorization header to AWS S3 where it is not needed and
   causes an error.
+- Improved error reporting
 
 ## [1.2] - 2024-03-25